src/requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
requests==2.31.0
PyYAML==6.0.1
